pyproject.toml
[project]
name = "matrix-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "docker>=7.1.0",
    "matrix-nio>=0.25.2",
]
